Squashed commit of the following:

commit 259736779d5b1b4a5edb353aa1d614d657b20df4
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Tue Aug 9 16:13:22 2022 +0800

    [ENH]:update chat view

commit fde4382dd785388785f571ea6adf82b610dd20f5
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Tue Aug 9 15:13:30 2022 +0800

    [ENH]:update cloud structure

+++ AgoraWidgets/Cloud/AgoraCloudWidgetModels.swift
@@ -186,7 +186,7 @@ struct AgoraCloudPublicCourseware: Convertable {
 struct AgoraCloudPublicConversion: Convertable {
     let type: String
     let preview: Bool
-    let scale: Int64
+    let scale: Float
     let outputFormat: String
     let canvasVersion: Bool?
 }
